Load achievement statistics as part of Achievement
This reduced the number of queries to exactly 1 and solves the N+1
problem (or at least what was similar to it).
